+++ src/Components/Search/FoodModal.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useContext } from 'react';
 import { FoodContext } from './FoodDataProvider';
-import { FoodDatabaseContext } from '../FoodLog/FoodDatabaseProvider'
 import { Modal, Table } from 'antd'
 import { url } from './FoodCard'
+
 import './Search.css'
 
 
@@ -34,13 +34,12 @@ import './Search.css'
 ];
 
 export const FoodModal = (props) => {
-    const { getFoodDatabase, setFoodDatabase, foodDatabase } = useContext(FoodDatabaseContext)
-    const { visible, setVisible, nutrition, foodName, addFood, usersFood } = useContext(FoodContext)
+    const { visible, setVisible, nutrition, foodName, addFood, getItem, item, getFoodDatabase } = useContext(FoodContext)
     let image = <img className="modalImage"src={url + foodName.image} alt="food" />
 
    useEffect(() => {
-        console.log(foodDatabase)
-   }, [nutrition])
+    
+   }, [item])
 
    const isVisible = () => {
        if(visible === true){
@@ -52,12 +51,7 @@ export const FoodModal = (props) => {
     let amount = nutrition.map(b => {
         return b.amount
     })
-    await usersFood().then(res => {
-        let a = res.filter(id => id.id === foodName.id)
-        return a[0]
-    }).then(res => {
-        if(res === undefined){
-            addFood({
+      addFood({
                 apiId: foodName.id,
                 mealId: 0,
                 userId: 1,
@@ -73,12 +67,18 @@ export const FoodModal = (props) => {
                 sodium: amount[7],
                 protein: amount[8]
             })
-        } 
-        getFoodDatabase()
-        setVisible(false)
+            
+              setVisible(false)
+            
+            
+          
+   }
+       
         
-    })
-}
+    
+
+
+ 
 
     return (
         <>
