+++ src/Components/Search/Search.css
@@ -53,3 +53,34 @@
 .nutritionTable {
     height: 32.8125rem;
 }
+
+.foodSavedContainer {
+    background-color: rgba(0,0,0,0.3);
+    min-width: 6.25rem;
+    max-height: 28.4375rem;
+}
+
+.foodSavedContainer h4 {
+    color: white;
+    margin-left: 15px;
+    margin-top: 5px;
+}
+
+.savedCard {
+   position: relative;
+   left: 5px;
+   height: 175px;
+}
+
+.foodDbSearch {
+  margin-left: 7px;
+}
+
+.foodDbCards {
+    overflow: auto;
+    max-height: 385px;
+}
+
+.addFoodIcon {
+    cursor: pointer;
+}
